便利ヘッダー削除
