+++ src/KingdomCraft/KingdomCraft/KingdomCraft.vcxproj.filters
@@ -1,7 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
-    <ClCompile Include="stdafx.cpp" />
     <ClCompile Include="Library\DX11Manager\DX11Manager.cpp">
       <Filter>Library</Filter>
     </ClCompile>
@@ -25,7 +24,6 @@
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="stdafx.h" />
     <ClInclude Include="Library\DX11Manager\DX11Manager.h">
       <Filter>Library</Filter>
     </ClInclude>
