+++ src/KingdomCraft/KingdomCraft/Main/SceneManager/Scene/GameScene/GameScene.cpp
@@ -14,6 +14,7 @@ SceneID GameScene::Control()
 {
 	return m_sceneID;
 }
+
 void GameScene::Draw()
 {
