+++ src/KingdomCraft/KingdomCraft/Main/SceneManager/SceneFactory/SceneFactory.cpp
@@ -1,8 +1,7 @@
 #include "SceneFactory.h"
-#include "..//..//..//stdafx.h"
 #include "..//Scene//GameScene//GameScene.h"
 #include "..//Scene//TitleScene//TitleScene.h"
-
+#include <Windows.h>
 Scene* SceneFactory::CreateScene(SceneID _sceneID)
 {
 	Scene* pScene = NULL;
