--- src/KingdomCraft/KingdomCraft/stdafx.cpp
@@ -1,7 +1,0 @@
-// stdafx.cpp : 標準インクルード DirectxLibrary.pch のみを
-// 含むソース ファイルは、プリコンパイル済みヘッダーになります。
-// stdafx.obj にはプリコンパイル済み型情報が含まれます。
-
-#include "stdafx.h"
-// TODO: このファイルではなく、STDAFX.H で必要な
-// 追加ヘッダーを参照してください。
